ThemeEditor: Extract most logic into a MainWidget class
